backend/requirements.txt
fastapi
uvicorn
qdrant-client
openai
python-dotenv
langchain-text-splitters
